Refactor reducers, split to separate files.
