+++ src/components/Game/Game.test.tsx
@@ -2,7 +2,6 @@ import React from 'react'
 import { shallow, mount, ReactWrapper } from 'enzyme'
 import { Game } from './'
 import { GameBoard } from './Board'
-import { Square } from './Board/Square/Square'
 import R from 'ramda'
 import { createStore } from 'redux'
 import { Provider } from 'react-redux'
