+++ src/helpers/reducer/history.ts
@@ -1,23 +1,8 @@
-import actions, { Action } from './actions'
+import { Board, History, HistoryElement, PLAYERS } from '../../types'
+import actions, { Action } from '../actions'
 import * as R from 'ramda'
-import { GameState, History, PLAYERS, HistoryElement, Board } from 'types'
-import { combineReducers } from 'redux'
-import { getCurrentBoardFromHistory } from './selectors'
+import { getCurrentBoardFromHistory } from '../selectors'
 import { createReducer } from 'redux-create-reducer'
-import { isIllegalMove } from '.'
-
-export const stepNumber: (
-  state: number | undefined,
-  action: Action,
-) => number = (state = 0, action) => {
-  if (action.type === actions.JUMP_TO_STEP) {
-    return action.payload
-  }
-  if (action.type === actions.CLICK_ON_SQUARE) {
-    return state + 1
-  }
-  return state
-}
 
 const getNextPlayer: (board: Board) => PLAYERS = R.pipe<
   Board,
@@ -36,13 +21,11 @@ const getNextPlayer: (board: Board) => PLAYERS = R.pipe<
     [R.equals(1), R.always(PLAYERS.O)],
   ]),
 )
-
 export const createNextSquares: (
   board: Board | undefined,
   action: Action,
 ) => Board = (board = Array(9).fill(null), action) =>
   R.update(action.payload, getNextPlayer(board), board)
-
 const createNextSquaresFromHistory: (
   history: History,
   action: Action,
@@ -50,24 +33,20 @@ const createNextSquaresFromHistory: (
   getCurrentBoardFromHistory,
   R.identity,
 ])
-
 const dropHistoryElements = R.useWith(
   R.flip<number, History, History>(R.take),
   [R.identity, R.pipe(R.prop('payload'), R.add(1))],
 )
-
 const getIsNext: (state: History) => boolean = R.pipe<
   History,
   HistoryElement,
   boolean
 >(R.last, R.prop('xIsNext'))
-
 const nextXIsNext: (state: History, action: Action) => boolean = R.pipe<
   History,
   boolean,
   boolean
 >(getIsNext, R.not)
-
 const appendMoveToHistory: (
   history: History,
   action: Action,
@@ -78,7 +57,6 @@ const appendMoveToHistory: (
   }) as any,
   R.identity,
 ])
-
 export const history: (
   state: History | undefined,
   action: Action,
@@ -89,23 +67,3 @@ export const history: (
     [actions.JUMP_TO_STEP]: dropHistoryElements,
   },
 )
-
-const unsafeReducer = combineReducers({ history, stepNumber })
-
-const returnState = R.nAry(1, R.identity)
-
-const illegalClickAction = (state: GameState, action: Action): boolean => {
-  if (action.type === actions.CLICK_ON_SQUARE) {
-    if (isIllegalMove(state, action)) {
-      return true
-    }
-  }
-  return false
-}
-
-const reducer: (
-  state: GameState | undefined,
-  action: Action,
-) => GameState = R.ifElse(illegalClickAction, returnState, unsafeReducer)
-
-export default reducer
